renamed module to better fit its actual contents.
